+++ thingiverse/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.0.6rc1.dev"
+__version__ = "0.0.7.dev"
 
 from thingiverse.types.exceptions import (
     ResourceNotFound, SearchException,
@@ -74,6 +74,18 @@ class Thingiverse(object):
 
     #     return res
 
+    def handle_response_error(self, response: requests.Response, message: Text):
+        """Simply checks for invalid response and raises errors if any"""
+        if response.status_code == 404:
+            exception = ResourceNotFound(message)
+            logging.exception(exception)
+            raise exception
+        elif response.status_code != 200:
+            exception = ThingiverseException(f"Unknown error: {response.json()}")
+            logging.error(f"Error requesting url:\n\t{response.url}")
+            logging.exception(exception)
+            raise exception
+
     def get_user_by_username(self, username: Text = "me") -> ThingiverseUser:
         """Get user personal information by `username`
 
@@ -94,15 +106,7 @@ class Thingiverse(object):
         logging.info(f"Making request to get user by username: {username}")
         res = requests.get(url)
 
-        if res.status_code == 404:
-            exception = ResourceNotFound(f"User with username {username} not found.")
-            logging.exception(exception)
-            raise exception
-        elif res.status_code != 200:
-            exception = ThingiverseException(f"Unknown error: {res.json()}")
-            logging.error(f"Error requesting url:\n\t{url}")
-            logging.exception(exception)
-            raise exception
+        self.handle_response_error(res, f"User with username {username} not found.")
 
         res_json = res.json()
         user_box = Box(res_json)
@@ -139,21 +143,282 @@ class Thingiverse(object):
         logging.info(f"Making request to get username data: {username}; term: {term}")
         res = requests.get(url)
 
-        if res.status_code == 404:
-            exception = ResourceNotFound(f"Username with username {username} not found.")
-            logging.exception(exception)
-            raise exception
-        elif res.status_code != 200:
-            exception = ThingiverseException(f"Unknown error: {res.json()}")
-            logging.error(f"Error requesting url:\n\t{url}")
-            logging.exception(exception)
-            raise exception
+        self.handle_response_error(res, f"Username with username {username} not found.")
 
         res_json = res.json()
         thing_box = Box(res_json)
         logging.info(f"Successfully retrieved user data with username {username}")
 
         return thing_box
+
+    def get_username_things(self, username: Text = None):
+        """Gets a list of things by `username`
+
+        Parameters
+        ----------
+        username: str
+            username to search things for
+
+        Usage:
+        >>> from thingiverse import Thingiverse
+        >>> thingy = Thingiverse(access_token="abc")
+        >>> thing = thingy.get_things_by_username(username="m4t0n")
+        """
+        if not username:
+            exception = UserException("'username' is required")
+            logging.exception(exception)
+            raise exception
+
+        access_token_param = f"?access_token={self.access_token}"
+        path = f"/users/{username}/things"
+
+        url = self.base_url + path + access_token_param
+        logging.info(f"Making request to get things for username: {username}")
+        res = requests.get(url)
+
+        self.handle_response_error(res, f"Username with username {username} not found.")
+
+        res_json = res.json()
+        # thing_box = Box(res_json)
+        logging.info(f"Successfully retrieved things for username {username}")
+
+        return res_json
+
+    def get_username_collected_things(self, username: Text = None):
+        """Gets a list of collected things by `username`
+
+        Parameters
+        ----------
+        username: str
+            username to search things for
+
+        Usage:
+        >>> from thingiverse import Thingiverse
+        >>> thingy = Thingiverse(access_token="abc")
+        >>> thing = thingy.get_username_collected_things(username="m4t0n")
+        """
+        if not username:
+            exception = UserException("'username' is required")
+            logging.exception(exception)
+            raise exception
+
+        access_token_param = f"?access_token={self.access_token}"
+        path = f"/users/{username}/all-collected-things"
+
+        url = self.base_url + path + access_token_param
+        logging.info(f"Making request to get collected things for username: {username}")
+        res = requests.get(url)
+
+        self.handle_response_error(res, f"Username with username {username} not found.")
+
+        res_json = res.json()
+        things_box = Box(res_json)
+
+        logging.info(f"Successfully retrieved collected things for username {username}")
+
+        return things_box
+
+    def get_username_unread_message_count(self, username: Text = None):
+        """Gets unread message count
+
+        Parameters
+        ----------
+        username: str
+            username to search things for
+
+        Usage:
+        >>> from thingiverse import Thingiverse
+        >>> thingy = Thingiverse(access_token="abc")
+        >>> thing = thingy.get_username_unread_message_count()
+        """
+
+        access_token_param = f"?access_token={self.access_token}"
+        path = f"/users/{username}/unread-message-count"
+
+        url = self.base_url + path + access_token_param
+        logging.info(f"Making request to get message count for username {username}")
+        res = requests.get(url)
+
+        error_msg = f"Something went wrong while fetching message count for username {username}."
+        self.handle_response_error(res, error_msg)
+
+        res_json = res.json()
+        # message_count_box = Box(res_json)
+        logging.info(f"Successfully retrieved message count for username {username}")
+        print("Unread messages")
+        print(res_json)
+        return res_json
+
+    def get_username_favorites(self, username: Text = None):
+        """Gets a list of favorites by `username`
+
+        Parameters
+        ----------
+        username: str
+            username to search favorites for
+
+        Usage:
+        >>> from thingiverse import Thingiverse
+        >>> thingy = Thingiverse(access_token="abc")
+        >>> thing = thingy.get_username_favorites(username="m4t0n")
+        """
+        if not username:
+            exception = UserException("'username' is required")
+            logging.exception(exception)
+            raise exception
+
+        access_token_param = f"?access_token={self.access_token}"
+        path = f"/users/{username}/favorites"
+
+        url = self.base_url + path + access_token_param
+        logging.info(f"Making request to get favorites for username: {username}")
+        res = requests.get(url)
+
+        self.handle_response_error(res, f"Username with username {username} not found.")
+
+        res_json = res.json()
+        # favorites_box = Box(res_json)
+        logging.info(f"Successfully retrieved favorites for username {username}")
+
+        return res_json
+
+    def get_username_likes(self, username: Text = None):
+        """Gets a list of likes by `username`
+
+        Parameters
+        ----------
+        username: str
+            username to search likes for
+
+        Usage:
+        >>> from thingiverse import Thingiverse
+        >>> thingy = Thingiverse(access_token="abc")
+        >>> thing = thingy.get_username_likes(username="m4t0n")
+        """
+        if not username:
+            exception = UserException("'username' is required")
+            logging.exception(exception)
+            raise exception
+
+        access_token_param = f"?access_token={self.access_token}"
+        path = f"/users/{username}/likes"
+
+        url = self.base_url + path + access_token_param
+        logging.info(f"Making request to get likes for username: {username}")
+        res = requests.get(url)
+
+        self.handle_response_error(res, f"Username with username {username} not found.")
+
+        res_json = res.json()
+        # likes_box = Box(res_json)
+        logging.info(f"Successfully retrieved likes for username {username}")
+
+        return res_json
+
+    def get_username_copies(self, username: Text = None):
+        """Gets a list of copies by `username`
+
+        Parameters
+        ----------
+        username: str
+            username to search copies for
+
+        Usage:
+        >>> from thingiverse import Thingiverse
+        >>> thingy = Thingiverse(access_token="abc")
+        >>> thing = thingy.get_username_copies(username="m4t0n")
+        """
+        if not username:
+            exception = UserException("'username' is required")
+            logging.exception(exception)
+            raise exception
+
+        access_token_param = f"?access_token={self.access_token}"
+        path = f"/users/{username}/copies"
+
+        url = self.base_url + path + access_token_param
+        logging.info(f"Making request to get copies for username: {username}")
+        res = requests.get(url)
+
+        self.handle_response_error(res, f"Username with username {username} not found.")
+
+        res_json = res.json()
+        # copies_box = Box(res_json)
+        logging.info(f"Successfully retrieved copies for username {username}")
+
+        return res_json
+
+    def get_username_collections(self, username: Text = None, paginate: bool = True):
+        """Gets a list of copies by `username`
+
+        Parameters
+        ----------
+        username: str
+            username to search copies for
+
+        paginate: bool, default=True
+            return a paginated response, if `False`, a full list will be returned
+
+        Usage:
+        >>> from thingiverse import Thingiverse
+        >>> thingy = Thingiverse(access_token="abc")
+        >>> thing = thingy.get_username_collections(username="m4t0n", paginate=False)
+        """
+        if not username:
+            exception = UserException("'username' is required")
+            logging.exception(exception)
+            raise exception
+
+        access_token_param = f"?access_token={self.access_token}"
+        path = f"/users/{username}/collections"
+
+        if paginate:
+            path += "/all"
+
+        url = self.base_url + path + access_token_param
+        logging.info(f"Making request to get collections for username: {username}")
+        res = requests.get(url)
+
+        self.handle_response_error(res, f"Username with username {username} not found.")
+
+        res_json = res.json()
+        # collections_box = Box(res_json)
+        logging.info(f"Successfully retrieved collections for username {username}")
+
+        return res_json
+
+    def get_username_downloads(self, username: Text = None):
+        """Gets a list of downloads by `username`
+
+        Parameters
+        ----------
+        username: str
+            username to search downloads for
+
+        Usage:
+        >>> from thingiverse import Thingiverse
+        >>> thingy = Thingiverse(access_token="abc")
+        >>> thing = thingy.get_username_downloads(username="m4t0n")
+        """
+        if not username:
+            exception = UserException("'username' is required")
+            logging.exception(exception)
+            raise exception
+
+        access_token_param = f"?access_token={self.access_token}"
+        path = f"/users/{username}/downloads"
+
+        url = self.base_url + path + access_token_param
+        logging.info(f"Making request to get copies for username: {username}")
+        res = requests.get(url)
+
+        self.handle_response_error(res, f"Username with username {username} not found.")
+
+        res_json = res.json()
+        # downloads_box = Box(res_json)
+        logging.info(f"Successfully retrieved copies for username {username}")
+
+        return res_json
 
     def get_thing_by_id(self, thing_id: int = None):
         """Gets a thing by `thing_id`
@@ -181,15 +446,7 @@ class Thingiverse(object):
         logging.info(f"Making request to get thing by id: {thing_id}")
         res = requests.get(url)
 
-        if res.status_code == 404:
-            exception = ResourceNotFound(f"Thing with id {thing_id} not found.")
-            logging.exception(exception)
-            raise exception
-        elif res.status_code != 200:
-            exception = ThingiverseException(f"Unknown error: {res.json()}")
-            logging.error(f"Error requesting url:\n\t{url}")
-            logging.exception(exception)
-            raise exception
+        self.handle_response_error(res, f"Thing with id {thing_id} not found.")
 
         res_json = res.json()
         thing_box = Box(res_json)
@@ -253,15 +510,7 @@ class Thingiverse(object):
         logging.info(f"Making request to get thing by id: {thing_id}")
         res = requests.get(url)
 
-        if res.status_code == 404:
-            exception = ResourceNotFound(f"Thing with id {thing_id} not found.")
-            logging.exception(exception)
-            raise exception
-        elif res.status_code != 200:
-            exception = ThingiverseException(f"Unknown error: {res.json()}")
-            logging.error(f"Error requesting url:\n\t{url}")
-            logging.exception(exception)
-            raise exception
+        self.handle_response_error(res, f"Thing with id {thing_id} not found.")
 
         res_json = res.json()
         # images_box = Box(res_json)