+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+# 0.0.7.dev (2022-02-17)
+
+- `handle_response_error()` method added
+  - raises exception on error responses
+  - Model added `thingivers.types.users.ThingiverseUser`
+- `get_username_things()` method added (`/users/{username}/things`)
+  - Fetches user's list of things
+- get_username_favorites() method added (`/users/{username}/favorites`)
+  - Fetches user's list of favorites
+- get_username_likes() method added (`/users/{username}/likes`)
+  - Fetches user's list of likes
+- `get_username_copies()` method added (`/users/{username}/copies`)
+  - Fetches user's list of copies
+- `get_username_collections()` method added (`/users/{username}/collections`)
+  - Fetches user's list of collections
+- `get_username_downloads()` method added (`/users/{username}/downloads`)
+  - Fetches user's list of downloads
+- `get_username_collected_things()` method added (`/users/{username}/all-collected-things`)
+  - Fetches a list of collected things by a given username
+- `get_username_unread_message_count()` method added (`/users/unread-message-count`)
+  - Fetches a count of unread messages
+
 # 0.0.6rc1.dev (2022-02-14)
 
 - `.dev` versioning started :)